requirements.txt
python-dotenv
mysql-connector-python
pymysql
scikit-learn
# -e .